Migrated project to TypeScript
