+++ tailwind.config.ts
@@ -1,4 +1,5 @@
-/** @type {import('tailwindcss').Config} */
+import type { Config } from "tailwindcss";
+
 export default {
   content: [
     "./pages/**/*.{js,ts,jsx,tsx,mdx}",
@@ -14,4 +15,4 @@ export default {
     },
   },
   plugins: [],
-};
+} satisfies Config;
